+++ 3D Engine/ECS/Entity.h
@@ -4,7 +4,7 @@
 class Entity
 {
 public:
-	static constexpr uint64 InvalidID = std::numeric_limits<uint64>::max();
+	static constexpr uint32 InvalidID = std::numeric_limits<uint32>::max();
 
 	Entity();
 
@@ -23,14 +23,14 @@ public:
 		return EntityID < Entity.GetEntityID();
 	}
 
-	uint64 GetEntityID() const;
+	uint32 GetEntityID() const;
 
 	explicit operator bool() const;
 
 private:
 	friend class EntityManager; // So that only the entity manager can assign entity ids.
 
-	uint64 EntityID;
+	uint32 EntityID;
 
-	Entity(uint64 EntityID);
+	Entity(uint32 EntityID);
 };
